Apply suggestions after review
